Externalized main word list
Moved the main word list to it's own JSON data file.  Left a smaller dictionary in the data model class file in case there are issues loading the data file.  The app now uses three sources for word lists - the static list in the class, the JSON data file which acts as the main store, and the external API - which could extend the functionality beyond what's available locally.  Cool.