CSS 7-1 FileSystem
I'm struggling to find a good way to apply Sass, so I did a bunch of reading.  It looks like it would be more meaningful on larger projects, but my reading introduced me to the 7-1 filesystem.  This looks like a good way to begin applying Sass in a systematic way.

The exercise forced me to look at my CSS as a whole and I found a couple items that were poorly written.

In addition to this, it would be a good idea to practice BEM.

This sets me up certainly to switch to Sass includes, partials, variables, nesting, and possibly mixins or extends/inheritance for similar items (like the monitors).

Specifically I can revisit adding Sass variables for the world sprite.  Testing seems to indicate hard coding these values helps with the flicker, using variables in the preprocessor should give me the best of both worlds.